Rename : 술 도감 패키지 네이밍 변경
1. spirits_collection -> liquor_collection
