+++ app/src/main/res/layout/fragment_liquor_collection.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.spirits_collection.LiquorCollectionFragment">
+    tools:context=".ui.liquor_collection.LiquorCollectionFragment">
 
     <androidx.compose.ui.platform.ComposeView
         android:id="@+id/compose_view"
